Moves ascii drawing into its own library
